+++ gen-bindings.sh
@@ -4,6 +4,6 @@ exec bindgen cpp/src/hltas.hpp \
 	-o hltas-cpp-bridge/src/hltas_cpp.rs \
 	--whitelist-function 'hltas_.*' \
 	--whitelist-type 'HLTAS::ErrorDescription' \
-	--rustified-enum 'HLTAS::StrafeType|StrafeDir|ButtonState|Button|ErrorCode' \
+	--rustified-enum 'HLTAS::StrafeType|StrafeDir|ButtonState|Button|ErrorCode|StrafingAlgorithm|ConstraintsType' \
 	--disable-name-namespacing \
 	-- -std=c++14 -I/usr/bin/../lib64/gcc/x86_64-pc-linux-gnu/9.2.0/include