+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## [Unreleased]
 ### Added
 - `Properties::hlstrafe_version`.
+- `Line::VectorialStrafing` and `Line::VectorialStrafingConstraints`.
 
 [Unreleased]: https://github.com/HLTAS/hltas/compare/v0.1.0...HEAD